requirements.txt
# Core
anthropic>=0.40.0
langgraph>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Market Data
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0

# ML / HMM
hmmlearn>=0.3.0
scikit-learn>=1.3.0

# UI
streamlit>=1.35.0
plotly>=5.20.0

# HTTP (for FRED)
requests>=2.31.0